Use 'reminder on' and 'reminder off' instead of 'start' and 'stop'
