+++ custom/ewsghana/handlers/help.py
@@ -1,6 +1,7 @@
 from corehq.apps.products.models import SQLProduct
-from custom.ewsghana.handlers import HELP_TEXT
+from custom.ewsghana.handlers import HELP_TEXT, DEACTIVATE_REMINDERS, REACTIVATE_REMINDERS
 from custom.ewsghana.handlers.keyword import KeywordHandler
+from custom.ewsghana.utils import user_needs_reminders
 
 
 class HelpHandler(KeywordHandler):
@@ -13,10 +14,11 @@ class HelpHandler(KeywordHandler):
         if topic == 'stock':
             self.respond("Please send your receipts in the format "
                          "' <Commodity code> <stock on hand > . <quantity received>'")
-        elif topic == 'stop':
-            self.respond("Text 'stop' to stop receiving text message reminders.")
-        elif topic == 'start':
-            self.respond("Text 'start' to get text message reminders every week to submit your stock reports.")
+        elif topic == 'reminder':
+            if user_needs_reminders(self.user):
+                self.respond(DEACTIVATE_REMINDERS)
+            else:
+                self.respond(REACTIVATE_REMINDERS)
         elif 'code' in topic:
             codes = SQLProduct.by_domain(self.domain).order_by('code').values_list('code', flat=True)
             self.respond("Available commodity codes: %(codes)s", codes=", ".join(codes))
